+++ srcs/controllers/index.tsx
@@ -4,7 +4,6 @@ export { default as addUser2Team } from './controller.addUser2Team';
 export { default as addUser2WaitList } from './controller.addUser2Waitlist';
 export { default as getTeam } from './controller.getTeam';
 export { default as removeUser2WaitList } from './controller.removeUser2WaitList';
-export { default as createGitRepo } from './controller.createGitRepo';
 export { default as getWaitlist } from './controller.getWaitlist';
 export { default as inviteToRepo } from './controller.inviteToRepo';
 export { default as updateTeam } from './controller.updateTeam';
